Revert "Ball colors etc"
This reverts commit 884b22e6391d4883697653647b966213122ac15f.

+++ labb4-pool/src/labb4/Hole.java
@@ -3,9 +3,9 @@ package labb4;
 public class Hole {
     public final int row;
     public final int col;
-    public final int radius;
+    public final double radius;
 
-    public Hole(int row, int col, int radius) {
+    public Hole(int row, int col, double radius) {
         this.row = row;
         this.col = col;
         this.radius = radius;
